Rewire quit command to close window
