+++ browser/src/Editor/OniEditor/OniEditor.tsx
@@ -14,6 +14,8 @@ import * as types from "vscode-languageserver-types"
 import * as Oni from "oni-api"
 import { IEvent } from "oni-types"
 
+import { remote } from "electron"
+
 import * as Log from "./../../Log"
 
 import { PluginManager } from "./../../Plugins/PluginManager"
@@ -136,6 +138,18 @@ export class OniEditor implements IEditor {
             this._tokenColors,
             this._workspace,
         )
+
+        this._neovimEditor.onNeovimQuit.subscribe(() => {
+            const isSplitModeOni = this._configuration.getValue("editor.split.mode") === "oni"
+
+            if (!this._configuration.getValue("debug.persistOnNeovimExit") && !isSplitModeOni) {
+                remote.getCurrentWindow().close()
+            } else if (isSplitModeOni) {
+                const handle = windowManager.getSplitHandle(this)
+                handle.close()
+                editorManager.unregisterEditor(this)
+            }
+        })
 
         this._neovimEditor.bufferLayers.addBufferLayer("*", buf =>
             wrapReactComponentWithLayer("oni.layer.scrollbar", <BufferScrollBarContainer />),
